Broken positioning correction element dependent on theme
Making the element become independent positioning of the parent
elements.
If a preceding element receiving the float property, the next element is
affected, this was the case. Isolated the element in a div.
If the test Storefront Tema, see the broken and totally misaligned code.
But with this code element is independent, obtaining full width without
affecting the posicinamento of elements around