Remove obsolete options
On the profile page, the titles and slugs of the tabs get overridden by
options which are not anymore manageable:
user_page_title_questions
user_page_slug_questions
user_page_title_answers
user_page_slug_answers

After upgrading, the options still exist in the Wordpress option table,
so the default ones set on the profile page (and translated) never get
displayed.

The easiest solution is to delete the obsolete options upon plugin
activation. A better solution would we an upgrade class when the
version of the plugin is changing, but this would be much much more
work.