ap_total_flagged_count always return zero
The query in ap_total_posts_count works wrong and groups the responses
by post_status. Since flag is not a post_status but a meta, the array
returned by ap_total_posts_count will never get an index of flag. If we
query them directly, the total returned by the function is then right.